registration extensions refactored
