+++ apps/InMeal.Api/RegistrationExtensions/ApplicationServiceRegistrationExtensions.cs
@@ -2,9 +2,6 @@
 using InMeal.Api.Features.Ingredients;
 using InMeal.Api.Features.Recipes;
 using InMeal.Api.Features.Upcoming;
-using InMeal.Infrastructure.External.GenerativeRecipeImages;
-using InMeal.Infrastructure.QueryServices;
-using InMeal.Infrastructure.Repositories;
 
 namespace InMeal.Api.RegistrationExtensions;
 
@@ -22,28 +19,7 @@ public static class ApplicationServiceRegistrationExtensions
             .As<ICancellationTokenAccessor>()
             .InstancePerLifetimeScope();
 
-        return containerBuilder
-            .RegisterQueryServices()
-            .RegisterRepositories()
-            .RegisterManagersAndServices()
-            .RegisterExternalServices();
-    }
-
-    private static ContainerBuilder RegisterQueryServices(this ContainerBuilder containerBuilder)
-    {
-        containerBuilder.RegisterType<AsyncRecipeQueryService>().AsImplementedInterfaces().InstancePerDependency();
-        containerBuilder.RegisterType<AsyncRecipeCategoryQueryService>().AsImplementedInterfaces().InstancePerDependency();
-        containerBuilder.RegisterType<AsyncRecipeIngredientQueryService>().AsImplementedInterfaces().InstancePerDependency();
-        
-        return containerBuilder;
-    }
-    
-    private static ContainerBuilder RegisterRepositories(this ContainerBuilder containerBuilder)
-    {
-        containerBuilder.RegisterType<AsyncIngredientRepository>().AsImplementedInterfaces().InstancePerDependency();
-        containerBuilder.RegisterType<AsyncRecipeRepository>().AsImplementedInterfaces().InstancePerDependency();
-        
-        return containerBuilder;
+        return containerBuilder.RegisterManagersAndServices();
     }
     
     private static ContainerBuilder RegisterManagersAndServices(this ContainerBuilder containerBuilder)
@@ -52,13 +28,6 @@ public static class ApplicationServiceRegistrationExtensions
         containerBuilder.RegisterType<RecipeManager>().AsImplementedInterfaces().InstancePerDependency();
         containerBuilder.RegisterType<IngredientsManager>().AsImplementedInterfaces().InstancePerDependency();
         
-        return containerBuilder;
-    }
-
-    private static ContainerBuilder RegisterExternalServices(this ContainerBuilder containerBuilder)
-    {
-        containerBuilder.RegisterType<GenerativeRecipeImages>().AsImplementedInterfaces().InstancePerDependency();
-
         return containerBuilder;
     }
 }
